converted total distance calculation from double to long long

+++ include/TSP.h
@@ -15,7 +15,7 @@ private:
     static Graph<int> getMST(const Graph<int>& graph);
     static void dfsTraversal(Vertex<int>* current, std::vector<Vertex<int>*>& path);
     static void traverseMST(const Graph<int>& graph, Vertex<int>* start, std::vector<Vertex<int>*>& path);
-    double pathDistance(const std::vector<Vertex<int>*>& path);
+    long long pathDistance(const std::vector<Vertex<int>*>& path);
   
 public:
     explicit TSP(const Graph<int> &graph);
